% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAngles.R
\name{convertAngles}
\alias{convertAngles}
\title{Convert alongship and athwartship split beam angles}
\usage{
convertAngles(ekraw)
}
\arguments{
\item{ekraw}{Imported EK60 data, as returned by \code{read.EK60_raw}.}
}
\value{
A four dimensions array [depth samples, pings, transceivers, 2]
with mechanical angles.
}
\description{
Convert angles in imported EK60 raw files from electrical to mechanical.
}
\examples{
if(interactive()){
ek <- read.EK60_raw("D20130504-T083828.raw", parseNMEA = TRUE, angles = TRUE)
angles <- convertAngles(ek)
dim(angles)
} 
}
\author{
Héctor Villalobos
}
