% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPower.R
\name{convertPower}
\alias{convertPower}
\title{Convert Received Power to Sv or TS}
\usage{
convertPower(ekraw, frequency = NULL, output = "Sv", saCorr)
}
\arguments{
\item{ekraw}{Imported EK60 data, as returned by \code{read.EK60_raw}.}

\item{frequency}{An integer corresponding to index of the transceiver
in the data to convert.}

\item{output}{A string indicating whether the power should be converted to 
acoustic backscattering strength (``Sv'') or to target strength (``TS'').}

\item{saCorr}{sa correction value from calibration}
}
\value{
A two dimensions array [depth samples, pings] with the desired 
output.
}
\description{
Convert acoustic power from imported EK60 raw files to Sv or TS data.
}
\details{
While the function operates on the EK60 data, it should be preferable 
used through \code{ek2echogram}, which will convert the ek data to echogram.
}
\examples{
if(interactive()){
ek <- read.EK60_raw("D20130504-T083828.raw", parseNMEA = TRUE, angles = TRUE)

Sv <- convertPower(ek, output = "Sv")
image(Sv)
}
}
\seealso{
\code{ek2echogram}.
}
\author{
Héctor Villalobos.
}
