% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgTime.R
\name{dgTime}
\alias{dgTime}
\title{Get datagram time from imported EK* raw files.}
\usage{
dgTime(raw, ini)
}
\arguments{
\item{raw}{An imported EK* raw file}

\item{ini}{Initial Byte with time data}
}
\value{
datagram time in POSIXct format
}
\description{
Function get.dgTime() is not meant to be run by the user. it 
is called by other functions.
}
\details{
The time is cpu time, and the time zone is assumed to be UTC
which may not be the case.
}
\examples{
 
 
}
