\name{echo.noise}
\alias{echo.noise}
\docType{data}
\title{
Sample echogram data (120 kHz)
}
\description{
Echogram for background noise estimation.
}
\usage{data("echo.noise")}

\details{
For this echogram, recording range (500 m) has been set above the sea bottom, and there is no evident presence of biological scatteres in order to facilitate the background noise estimation.
}
\examples{
data("echo.noise")
echogram(echo.noise)
}
\keyword{datasets}
