% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ek2echogram.R
\name{ek2echogram}
\alias{ek2echogram}
\title{Convert EK60 data to echogram}
\usage{
ek2echogram(
  ekraw,
  frequency = 1,
  data = "Sv",
  environment = NULL,
  calibration = NULL
)
}
\arguments{
\item{ekraw}{Imported EK60 data, as returned by \code{read.EK60_raw}.}

\item{frequency}{An integer corresponding to index of the transceiver
in the data to be converted.}

\item{data}{A string indicating the desired data: acoustic backscattering 
strength (``Sv'') or target strength (``TS'').}

\item{environment}{An optional list with two elements: ``soundVelocity'' and 
``absorptionCoeff'' calculated according to sea temperature and salinity
measured during data collection.}

\item{calibration}{A list as above, with ``gain'' and ``saCorr'' obtained 
from the echosounder calibration.}
}
\value{
A vector with sample range (m).
}
\description{
Convert imported EK60 raw data to class ``echogram''.
}
\details{
This function calls the \code{sampleRange} and \code{convertPower} 
functions to produce an object of  \code{class} ``echogram''.
}
\examples{
if(interactive()){
ek <- read.EK60_raw("D20130504-T083828.raw", parseNMEA = TRUE, angles = TRUE)

eco <- ek2echogram(ek, frequency = 1, data = "Sv")
echogram(eco)
}
}
\seealso{
\code{read.echogram}.
}
\author{
Héctor Villalobos.
}
