% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CON0.R
\name{get_CON0}
\alias{get_CON0}
\title{Get CON0 datagram from EK60 raw files}
\usage{
get_CON0(raw)
}
\arguments{
\item{raw}{A raw vector imported via \code{read.EK_raw} or an EK60 raw file name.}
}
\value{
A list with two data frames: Header, containing survey-, transect- 
and sounder names, software version, and number of transceivers; Transceiver,
with channel ID, beamtypes, frequency, gain, equivalent beam angle, etc. for 
each transceiver.
}
\description{
Read the echosounder configuration information stored in CON0 datagram.
}
\details{
While it can be used for reading the configuration data from an 
EK60 raw file, this function is not meant to be called directly by the user.
}
\examples{
if(interactive()){
config <- get_CON0("D20130504-T083828.raw")
config
}
}
\author{
Héctor Villalobos.
}
