% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dgIdx.R
\name{get_dgIdx}
\alias{get_dgIdx}
\title{Get datagram indices from EK* raw files}
\usage{
get_dgIdx(raw)
}
\arguments{
\item{raw}{A raw vector imported via \code{read.EK_raw} or an EK* raw file name}
}
\value{
A data frame with datagram types (CON0, NME0, RAW0, etc.) and lengths, 
        and most important, the indices for extracting each datagram.
}
\description{
Find the datagram types and lengths, and their respective indices in EK* raw files
}
\details{
In EK* raw files, the first 4 bytes before every datagram contain its
length (dgLen), which is repeated after the datagram data, before the length 
of the next datagram. So, for skipping from one datagram to the next, it takes
4 + dgLen + 4 + 1.  
Within each datagram, the first 4 bytes give its name (dgType: CON0, NME0, RAW0, etc.),
followed by the time in the next 8 bytes, and then the datagram data according to its type. 
The output of this fuction is key for extracting configuration data (header and 
transceiver info), nmea sentences, and acoustic raw data (received power and angles).
}
\examples{
 \dontrun{
 dgIdx <- dgIndices(ekraw)
 }
 
}
\author{
Héctor Villalobos
}
