\name{mergeSvmat}
\alias{mergeSvmat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Merge inequal Sv data matrices
}
\description{
Internal function called by \code{read.echogram} and \code{merge.echogram}.
}
\usage{
mergeSvmat(m1, m2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{m1}{
First Sv data matrix.
}
  \item{m2}{
Second Sv data matrix.
}
}

\author{
Héctor Villalobos
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ array }
