% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.nmea.R
\name{parse.nmea}
\alias{parse.nmea}
\title{Parse NMEA sentences}
\usage{
parse.nmea(nmea, sentence = NULL, returnAll = FALSE)
}
\arguments{
\item{nmea}{A data frame with CPU time and corresponding NMEA sentences
as returned by \code{get_NME0}.}

\item{sentence}{A string identifying the NMEA sentence to process. It
can be one of "GPRMC", "GPGGA", "INGGA", or "GPGLL".}

\item{returnAll}{logical. If \code{TRUE}, all data found in the selected
sentence is returned.}
}
\value{
A data frame with cpu time, gps time, miliseconds (already added
in gps time when the chossen sentence is "GPRMC"), longitude, and
latitude. Additionally, vessel speed and bearing are returned if found in
the data.
}
\description{
Find and parse GPS data from NMEA sentences.
}
\details{
This function looks first for the "GPRMC" sentence, if not found  
it tries then with "GPGGA", "INGGA", or "GPGLL", and process the one 
more information (more GPS fixes). Be ware that sometimes strings may be
corrupted, so it is possible to choose manually the sentence to process.
This is why \code{get_NME0} does not parse the NMEA strings by default. 
When one of "GPVTG", "INVTG", "GNVTG", or "IIVTG" sentences is found, vessel 
speed and bearing are also parsed and returned. 
Differences in time.cpu and gps.time may reveal wrong clock settings in the 
data aquisition computer.
}
\examples{
if(interactive()){
ek <- read.EK60_raw("D20130504-T083828.raw", angles = FALSE)

gps <- parse.nmea(ek$nmea)
head(gps)
}
}
\references{
NMEA sentences structure can be seen in: 
\url{https://gpsd.gitlab.io/gpsd/NMEA.html}
}
\seealso{
\code{get_NME0}.
}
\author{
Héctor Villalobos.
}
