% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.EK_raw.R
\name{read.EK_raw}
\alias{read.EK_raw}
\title{Read EK* raw files from Simrad echosounders}
\usage{
read.EK_raw(file)
}
\arguments{
\item{file}{EK* raw file name}
}
\value{
A vector of class ``raw''.
}
\description{
This function will read the whole Simrad EK* raw file as raw bytes.
}
\details{
This function is not intended to be called directly by the user, as  
it returns hex digits that should be first interpreted to be useful. Instead, 
\code{read.EK_raw} is called by other echogram functions that also convert 
the data to the appropriate type.
}
\author{
Héctor Villalobos
}
