% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRange.R
\name{sampleRange}
\alias{sampleRange}
\title{Determine acoustic samples depth range}
\usage{
sampleRange(ekraw, frequency = NULL)
}
\arguments{
\item{ekraw}{Imported EK60 data, as returned by \code{read.EK60_raw}.}

\item{frequency}{An integer corresponding to index of the transceiver
in the data for which to obtain the sample range.}
}
\value{
A vector with sample range (m).
}
\description{
Calculate sample range according to sound speed and sample interval for imported EK60 raw files.
}
\details{
While the function operates on the EK60 data, it should be preferable 
used through \code{ek2echogram}, which will convert the ek data to echogram.
}
\examples{
if(interactive()){
ek <- read.EK60_raw("D20130504-T083828.raw", parseNMEA = TRUE, angles = TRUE)

R <- sampleRange(ek, frequency = 1)
R
}
}
\seealso{
\code{ek2echogram}.
}
\author{
Héctor Villalobos.
}
