% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcvrConf.R
\name{xcvrConf}
\alias{xcvrConf}
\title{Extract transceiver configuration from imported EK* raw files.}
\usage{
xcvrConf(raw, ini)
}
\arguments{
\item{raw}{A raw vector imported via \code{read.EK_raw}.}

\item{ini}{Initial byte with transceiver data}
}
\value{
a data frame with transceiver configuration info.
}
\description{
Extract datagram time and convert it to POSIXct.
}
\details{
This function should not be called directly by the user.
}
\author{
Héctor Villalobos
}
