% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CON0.R
\name{get_CON0}
\alias{get_CON0}
\title{Get CON0 datagram from EK60 raw files}
\usage{
get_CON0(raw)
}
\arguments{
\item{raw}{An imported EK60 raw file or a raw file name}
}
\value{
A list with header and transceiver configuration
}
\description{
Read the echosounder configuration information stored in CON0 datagram.
}
\details{
Not to be called directly by the user.
}
\examples{
fn <- system.file("extdata", "demo-D20130504-T083828.raw", package = "echogram")
config <- get_CON0(fn)
config

}
\author{
Héctor Villalobos
}
