% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NME0.R
\name{get_NME0}
\alias{get_NME0}
\title{Get NME0 datagrams from imported EK* raw files.}
\usage{
get_NME0(raw)
}
\arguments{
\item{raw}{An imported EK* raw file or a raw file name}
}
\value{
A data frame with CPU time and corresponding NMEA data
}
\description{
Read the NMEA sentences stored in NME0 datagrams.
}
\details{
For parsing NMEA sentences, \code{parse.nmea()} function
should be used.
}
\examples{
fn <- system.file("extdata", "demo-D20130504-T083828.raw", package = "echogram")
nmea <- get_NME0(fn)
head(nmea)
 
}
\seealso{
\code{parse.nmea}
}
\author{
Héctor Villalobos
}
