% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NME0.R
\name{get_NME0}
\alias{get_NME0}
\title{Get NMEA sentences from NME0 datagrams in imported EK* raw files.}
\usage{
get_NME0(raw)
}
\arguments{
\item{raw}{An imported EK* raw file or a raw file name}
}
\value{
A data frame with CPU time and corresponding NMEA data
}
\description{
Get NMEA sentences from NME0 datagrams in imported EK* raw files.
}
\examples{
 \dontrun{
 nmea <- get.NME0(ekraw)
 }
 
}
