% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RAW0.R
\name{get_RAW0}
\alias{get_RAW0}
\title{Get RAW0 datagrams from imported EK60 raw files.}
\usage{
get_RAW0(raw, angles)
}
\arguments{
\item{raw}{An imported EK60 raw file or a raw file name}

\item{angles}{Logical. If TRUE, and angle data is present, an array
containing alongship and atwarthship electrical angles is returned.}
}
\value{
A list with sample data information, and an array with received power.
}
\description{
Read the acoustic sample data stored in RAW0 datagrams.
}
\details{
Not to be called directly by the user.
}
\author{
Héctor Villalobos
}
