% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.nmea.R
\name{parse.nmea}
\alias{parse.nmea}
\title{Parse NMEA sentences}
\usage{
parse.nmea(nmea, sentence = NULL, returnAll = FALSE)
}
\arguments{
\item{nmea}{A data frame with CPU time and corresponding NMEA sentences
as returned by \code{get_NME0}.}

\item{sentence}{A string identifying the NMEA sentence to process. It
can be one of "GPGGA", "INGGA", "GPGLL", "GPRMC".}

\item{returnAll}{logical. If \code{TRUE}, all data found in the selected
sentence is returned.}
}
\value{
A data frame with cpu time, gps time, miliseconds, longitude, and
latitude. Also, vessel speed and bearing if found in the data.
}
\description{
Find and parse GPS data from NMEA sentences.
}
\details{
This function identifies among the sentences found in the data, 
the one with more information (more GPS fixes). However, sometimes they 
may be corrupted, so the user can choose a different sentence to process.
This is why \code{get_NME0} does not parse the NMEA strings by default. 
If sentences: "GPVTG", "INVTG", "GNVTG", or "IIVTG" are found, vessel 
speed and bearing are also parsed and returned.
}
\examples{
if(interactive()){
ek <- read.EK60_raw("D20130504-T083828.raw", angles = FALSE)

gps <- parse.nmea(ek$nmea)
head(gps)
}
}
\references{
NMEA sentences structure can be seen in: 
\url{https://gpsd.gitlab.io/gpsd/NMEA.html}
}
\seealso{
\code{get_NME0}.
}
\author{
Héctor Villalobos.
}
