% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.EK60_raw.R
\name{read.EK60_raw}
\alias{read.EK60_raw}
\title{Read raw files from Simrad EK60 scientific echosounders}
\usage{
read.EK60_raw(file, parseNMEA = FALSE, angles = TRUE)
}
\arguments{
\item{file}{EK60 raw file name}

\item{parseNMEA}{logical. When false (the default) all the nmea sentences are
returned as is. If TRUE, the parse.nmea function try to parse the most appropriate.}
}
\value{
A list .
}
\description{
This function will read all data in EK60 raw files.
}
\details{
This function first calls \code{read.EK_raw}, and \code{get_dgIdx} functions.
With the raw vector and datagram indices returned, it finds the and extracts the 
configuration (CON0), GPS positions (NMEO) and acosutic data (RAW0) datagrams.
}
\author{
Héctor Villalobos
}
