% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.EK_bot.R
\name{read.EK_bot}
\alias{read.EK_bot}
\title{Read EK60 bot files from Simrad echosounders}
\usage{
read.EK_bot(file)
}
\arguments{
\item{file}{EK60 bot file name}
}
\value{
A data frame with ping times and depth data for every transceiver in 
the file.
}
\description{
This function will read Simrad BOT0 datagrams with bottom depth data.
}
\details{
This function calls \code{read.EK_raw}, and \code{get_dgIdx} functions
to read the bot file, find the BOT0 datagrams, and then extract ping times and
depth(s) associated with the transceiver(s) in the file.
}
\examples{
fn <- system.file("extdata", "demo-D20130504-T083828.bot", package = "echogram")
depth <- read.EK_bot(fn)
head(depth)

}
\author{
Héctor Villalobos
}
