% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.EK_raw.R
\name{read.EK_raw}
\alias{read.EK_raw}
\title{Read EK* raw files from Simrad echosounders}
\usage{
read.EK_raw(file)
}
\arguments{
\item{file}{EK* raw file name.}
}
\value{
A vector of class ``raw''.
}
\description{
This function imports Simrad EK60 or EK80 raw files as raw bytes.
}
\details{
This function will return raw bytes as hex digits, therefore is not   
intended to be called directly by the user. Instead, \code{read.EK_raw} is 
called by other echogram's functions that also convert the various datagrams
in EK60 files to their appropriate type.
}
\seealso{
\code{read.EK60_raw}, \code{get_CON0}, \code{get_NME0}, and
\code{get_RAW0}.
}
\author{
Héctor Villalobos.
}
