% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcvrConf.R
\name{xcvrConf}
\alias{xcvrConf}
\title{Extract transceiver configuration from EK60 raw files}
\usage{
xcvrConf(raw, ini)
}
\arguments{
\item{raw}{A raw vector imported via \code{read.EK_raw}.}

\item{ini}{Initial byte with transceiver data}
}
\value{
a data frame with transceiver configuration: channel ID, beamtypes, 
frequency, gain, equivalent beam angle, etc. for each transceiver.
}
\description{
Find and extract transceiver configuration information from imported EK60 raw files.
}
\details{
The index (ini) comes from output of \code{get_dgIdx} 
function. This function should not be called directly by the user.
}
\author{
Héctor Villalobos
}
